pyproject.toml
[project]
name = "map-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.13.2",
    "aioredis==1.3.1",
    "aiosignal==1.4.0",
    "alembic==1.17.2",
    "amqp==5.3.1",
    "annotated-doc==0.0.4",
    "annotated-types==0.7.0",
    "anyio==4.11.0",
    "async-timeout==5.0.1",
    "asyncpg==0.30.0",
    "attrs==25.4.0",
    "bcrypt==4.0.1",
    "billiard==4.2.4",
    "black==25.11.0",
    "celery==5.6.0",
    "certifi==2025.11.12",
    "click==8.3.1",
    "click-didyoumean==0.3.1",
    "click-plugins==1.1.1.2",
    "click-repl==0.3.0",
    "colorama==0.4.6",
    "dnspython==2.8.0",
    "email-validator==2.3.0",
    "exceptiongroup==1.3.1",
    "fastapi==0.121.2",
    "fastapi-cache2==0.2.2",
    "fastapi-cli==0.0.16",
    "fastapi-cloud-cli==0.3.1",
    "frozenlist==1.8.0",
    "greenlet==3.2.4",
    "h11==0.16.0",
    "hiredis==3.3.1",
    "httpcore==1.0.9",
    "httptools==0.7.1",
    "httpx==0.28.1",
    "idna==3.11",
    "iniconfig==2.3.0",
    "jinja2==3.1.6",
    "kombu==5.6.1",
    "mako==1.3.10",
    "markdown-it-py==4.0.0",
    "markupsafe==3.0.3",
    "mdurl==0.1.2",
    "multidict==6.7.0",
    "mypy-extensions==1.1.0",
    "nodeenv==1.10.0",
    "packaging==25.0",
    "passlib==1.7.4",
    "pathspec==0.12.1",
    "pendulum==3.1.0",
    "pillow==12.0.0",
    "platformdirs==4.5.0",
    "pluggy==1.6.0",
    "prompt-toolkit==3.0.52",
    "propcache==0.4.1",
    "pydantic==2.12.4",
    "pydantic-core==2.41.5",
    "pydantic-settings==2.12.0",
    "pygments==2.19.2",
    "pyjwt==2.10.1",
    "pyright==1.1.408",
    "pytest==9.0.2",
    "pytest-asyncio==1.3.0",
    "pytest-dotenv==0.5.2",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.2.1",
    "python-multipart==0.0.20",
    "pytokens==0.3.0",
    "pyyaml==6.0.3",
    "redis==4.6.0",
    "rich==14.2.0",
    "rich-toolkit==0.15.1",
    "rignore==0.7.6",
    "ruff==0.14.13",
    "sentry-sdk==2.44.0",
    "shellingham==1.5.4",
    "six==1.17.0",
    "sniffio==1.3.1",
    "sqlalchemy==2.0.44",
    "starlette==0.49.3",
    "typer==0.20.0",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.2",
    "tzdata==2025.3",
    "tzlocal==5.3.1",
    "urllib3==2.5.0",
    "uv==0.11.3",
    "uvicorn==0.38.0",
    "uvloop==0.22.1",
    "vine==5.1.0",
    "watchfiles==1.1.1",
    "wcwidth==0.2.14",
    "websockets==15.0.1",
    "yarl==1.22.0",
]

[tool.ruff]
line-length = 80

[tool.ruff.format]
docstring-code-line-length = 60
